main/requirements.txt
tk
pillow
keras
numpy
tensorflow
matplotlib
scipy
opencv-python